Separate OrtDownloadProcessor into Impl
This to decouple the functionality from the
AntennaContext.
